Redone in Cmake
Set up Cmake file, still need to fix some issues
