+++ app/parser/CMakeLists.txt
@@ -152,6 +152,6 @@ endif()
 add_library(pbf_parser_generated ${LIB_TYPE} ${GENERATED_PARSER_FILES})
 target_include_directories(pbf_parser_generated PUBLIC ${PROTO_LIBRARIES})
 target_link_libraries(pbf_parser_generated libprotobuf)
-add_library(pbf_parser ${LIB_TYPE} "${CMAKE_CURRENT_LIST_DIR}/pbf-parser.h" "${CMAKE_CURRENT_LIST_DIR}/pbf-parser.cpp")
+add_library(pbf_parser ${LIB_TYPE} "${CMAKE_CURRENT_LIST_DIR}/pbf-parser.h")
 target_include_directories(pbf_parser PUBLIC ${PROTO_LIBRARIES} "${ZLIB_PATH}")
 target_link_libraries(pbf_parser pbf_parser_generated ${ZLIB})
